backend/requirements.txt
python-decouple
annotated-types==0.7.0
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.7.0
django-ninja==1.5.3
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
whitenoise==6.11.0
